店舗推薦ロジックを別クラスとして切り出し
